Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/connect-src-eventsource-blocked-expected.txt

CONSOLE MESSAGE: Refused to connect to http://127.0.0.1:8000/eventsource/resources/simple-event-stream.asis because it does not appear in the connect-src directive of the Content Security Policy.
Pass
